+++ src/main/java/chess/repository/UserDao.java
@@ -80,6 +80,6 @@ public class UserDao implements UserRepository {
     }
 
     private User createUser(final ResultSet resultSet) throws SQLException {
-        return new User((long) resultSet.getInt("user_id"), resultSet.getString("name"));
+        return new User(resultSet.getLong("user_id"), resultSet.getString("name"));
     }
 }
